Reduced arch-chroot usage
Getting Error: GDBus.Error:org.freedesktop.DBus.Error.ServiceUnknown: The name org.freedesktop.PackageKit was not provided by any .service files

I believe it's related to software-properties-common and is a false positive. But the issue is that it causes arch-chroot to abort and doesn't cleanly un-mount the rootfs. I also experienced this at other points. I tried just using the 1 chroot and running the install script once, that seems to work a lot better. A future solution could be the use of ansible etc.